+++ server/models/angular-cli-process-map.model.ts
@@ -1,0 +1,6 @@
+import { AngularCliProcess } from "./angular-cli-process.model";
+
+export interface AngularCliProcessMap {
+    [guid: string]: AngularCliProcess;
+  }
+
