+++ server/models/angular-cli-process-status.enum.ts
@@ -1,0 +1,6 @@
+export enum AngularCliProcessStatus {
+    working = 0,
+    done = 1,
+    error = 2
+  }
+
